COM-2877: move usefull docs section
